Add web interface for seeing updates
This adds sinatra to begin a simple HTTP endpoint that I can check for
updates. Eventually I expect the web app to become more featureful.

This assumes that each model implements a '#date' method and that the
activities are ordered in the database by recency, which is probably not
valid cases of backfilling.