Configure rspec via a "test" environment
Although 256070f added the default rspec boilerplate, wiring it up into
the project is not trivial. This adds rspec configuration for a "test"
environment as given by a series of ".env" files which override each
other in a manner useful for environment-specific configuration:

  .env (lowest precedence)

  |    (overridden by:)
  v

  .env.test (higher precedence) (e.g.)

This is only currently useful for database tasks so as to prevent
fixtures from messing around with development data. See the new
README.md for usage details.

There's still some work to be done around improving the test data
quality, and of course test coverage, but hopefully this is a good start
:D.