Re-enable Rake tasks after running them the first time
The sync jobs would never run a subsequent time, due to Rake having a
good memory about which tasks it has already run and not repeating any.
This is kind of a weird use of Rake, so maybe another approach is called
for. For now, though, we can just use Task#reenable.