Move processing logic for dropbox files out of Rakefile
This is a first stab at a refactor to move processing logic out of the
Rakefile. This will make it more testable.

The new AtomicDropboxFile mixin is a method which is responsible for the
logic of fetching files from dropbox and calling processing code for
files which have not yet been processed (according to a now-required
`file_name` on the model.)